KUDU-1973 Fix caching in MultiRaftManager.
Currently, if another TServer disappears or we stop having followers on it
for any consensus, and later we need batching for it again, the batcher
is not cached, and every Raft consensus gets a distinct batcher.
This defeats the whole purpose of the feature and loses the
performance improvements.
However, this bug still allows Raft to work correctly, which is why unit
tests are green even with --enable_multi_raft_heartbeat_batcher=1.

Also adding spaces to the multi_raft_heartbeat_window_ms flag description.